feat: full rewrite using `getAll` and `setAll` cookie methods (#1)
Re-implements the `createServerClient` and `createBrowserClient` to
solve a few important problems that were not solved with the
implementation previously:

**Cookie encoding**
Cookies can't contain just any characters -- they're a limited set of US
ASCII and `"` is not allowed. Internally the Supabase client uses JSON
to encode the session, so this must not be mapped directly to cookies.
In this implementation, cookies are encoded as Base64 URL and the
leading chunk contains the `base64-` prefix.

**Cookie chunk management**
There are a few state transitions when storage needs to be chunked into
multiple cookies which were not handled well. This is the reason why a
change of API is necessary from `get`, `set` and `remove` cookie access
methods to `getAll` and `setAll`. The library must be able to see all of
the cookies and manage deleting unnecessary chunks when that is
required.

**Server-side developer ergonomics**
In some SSR frameworks (notably middleware in NextJS) it can be very
tricky to set cookies on objects correctly. `createServerClient` only
calls the `setAll` cookie access method only once and only when
necessary. This reduces the number of bugs developers can introduce
themselves when using the server client.

For more details check the `docs/design.md` document which contains a
lot of details that can't easily be covered in the PR description.