ci: fix doubling of `rc` version identifier (#13)
When an RC is published on top of an existing RC, the `rc` identifier
doubles. Example from
[`@supabase/auth-js`](https://www.npmjs.com/package/@supabase/auth-js/v/2.64.3-rc.rc.rc.3)
which uses pretty much the same Github Actions workflow file.

The fix is to change the `git describe --tags --exclude` to a value that
actually excludes [tags that start with
`rc`](https://github.com/supabase/ssr/releases/tag/rc0.4.0-rc.2).