fix: allow use of `createBrowserClient` without `window` present (#20)
Allows calling `createBrowserClient` without throwing an error when not
in a browser context (i.e. `window` is missing).

This is especially useful when using in Next.js Pre-Rendering mode where
the client's non-`auth` namespaces could be useful, such as to load
public data from PostgREST.

If however the `auth` namespace is used, the `setAll` handler function
will throw an error indicating to the user that they're doing something
wrong. This shouldn't come up in Pre-Rendering mode ever.

See:
https://github.com/orgs/supabase/discussions/27037#discussioncomment-9852948