fix: remove optional dependencies (#41)
## What kind of change does this PR introduce?

This is causing an issue for a user when deploying to vercel as it's
trying to install this optional dependency. However it doesn't look like
this optional dependency is being used anywhere in this project.

## What is the current behavior?

Optional dependency is not being used.

## What is the new behavior?

Remove optional dependency.

## Additional context

Add any other context or screenshots.