feat: adds `cookies.encode` option allowing minimal cookie sizes (#126)
Adds an experimental option `encode` on the `cookies` object when using
`createBrowserClient()` and `createServerClient()`.

If this is set to `tokens-only` then only the user's access token and
refresh token will be encoded in the cookies, causing significant cookie
size savings, often greater than 50%. It utilizes [split session storage
in `auth-js`](https://github.com/supabase/supabase-js/pull/1545), with
some trade-offs such as the inability to access the `user` property on
the `supabase.auth.getSession()` object in the server. This wasn't
supposed to be done anyway, and `getClaims()` is a secure alternative
for it.